New Feature Embedding Strategy
With a few caveats - the numeric features are being ignored, the user embedding is averaged so that the rankers will accept them, and the listing embeddings are collapsed to 768-dimensional vectors using PCA. Probable a few other shortcomings since I wanted to run though the whole pipeline. The rankers have not been re-trained.